feat(interface): 后端 plugins 模块整体更名为 interface，并将 Admin/Models 端点迁入 interface/models_api
- 源码目录 negentropy/plugins/ 整体迁移至 negentropy/interface/，路由前缀 /plugins → /interface；logger 名同步
- 从 auth/api.py 摘出 6 条 Admin Models 路由及相关工具函数，独立为 interface/models_api.py，前缀 /interface/models/*，保留 admin role 校验（_vendor_config_to_dict / _mask_api_key / _sanitize_error / SUPPORTED_VENDOR_CONFIG_VENDORS 随同迁入）
- rbac.PERMISSIONS 新增 interface:read / interface:write；admin 角色透过 interface:* 通配保持访问
- engine/bootstrap.py 更新 import 与 /plugins → /interface 挂载守卫及日志文案
- /interface/stats 新增 models: {total, enabled, vendors} 聚合字段（VendorConfig 计数 + ModelConfigRecord.enabled 计数）
- 测试目录 tests/unit_tests/plugins → tests/unit_tests/interface，test_model_configs_api/test_admin_models_ping 更名并迁入新目录，URL 断言同步切换至 /interface/models/*
- knowledge/extraction.py 的模块引用随命名统一更新
