+++ apps/negentropy/src/negentropy/interface/__init__.py
@@ -1,0 +1,10 @@
+"""
+Interface 模块。
+
+提供 MCP Server、Skill、SubAgent、Models 的管理和权限控制功能。
+"""
+
+from .api import router
+from .models_api import router as models_router
+
+__all__ = ["router", "models_router"]
