--- apps/negentropy/src/negentropy/plugins/__init__.py
@@ -1,9 +1,0 @@
-"""
-Plugins 模块。
-
-提供 MCP Server、Skill、SubAgent 的管理和权限控制功能。
-"""
-
-from .api import router
-
-__all__ = ["router"]
